add predict route and decorator jwt

+++ create.sql
@@ -29,9 +29,9 @@ CREATE TABLE histories (
     id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
     name VARCHAR(255),
     user_id UUID REFERENCES users(id),
-    creativity_score INTEGER NOT NULL,
-    logical_score INTEGER NOT NULL,
-    complexity_score INTEGER NOT NULL,
-    level_score INTEGER NOT NULL,
+    creativity_score FLOAT NOT NULL,
+    logical_score FLOAT NOT NULL,
+    complexity_score FLOAT NOT NULL,
+    level_score FLOAT NOT NULL,
     created_at TIMESTAMP
 );
